scripts/Cluster Evaluation/requirements.txt
pandas>=2.0.0
requests>=2.28.0
openpyxl>=3.1.0
google-generativeai>=0.3.0
